+++ CHANGELOG.md
@@ -42,6 +42,8 @@ CLI command and its behaviour. There are no guarantees of stability for the
 
 ### Added
 
+- `--fallback-dot-license` option added to `annotate` command. (#)
+
 ### Changed
 
 ### Deprecated